+++ src/app/overview/deck/deck.component.css
@@ -6,7 +6,7 @@
   border-radius: 15px;
 
   background-color: white;
-  box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
+  /*box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;*/
 }
 
 p {
